+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/nik-rev/subdef/compare/v0.1.4...HEAD
 
+- Fixed attributes being inserted at the end instead of at the start
+
 ## [v0.1.4] - 2026-02-27
 
 [v0.1.4]: https://github.com/nik-rev/subdef/compare/v0.1.3...v0.1.4